Major cleanup of installation. Removed hard-coded directory paths. CMake does not try to compile wrappers automatically any more as this adds plenty of dependencies. Option added to allow building of wrappers (as has been done in the past). Fixing of wscripts to deal with changed siox-inst behavior.
Waf: Allow parallel installation of multiple wrappers (MPICH2 vs. OpenMPI).
Added devel/scripts/build-wrapper.sh script to simplify installation of wrappers.

+++ devel/scripts/build-all-wrappers.sh
@@ -1,25 +1,45 @@
 #!/bin/bash
 
-SIOX=/opt/siox/Inst/
-CLEAN="$1"
+SIOX="$1"
+CLEAN="$2"
 
-(
-ERROR=""
-for L in /opt/siox/siox/src/tools/siox-skeleton-builder/layers/*; do
+CUR_DIR="$PWD"
+DIR=$(dirname $0)
+cd $DIR/../../
+
+if [[ "$SIOX" == ""  || ! -d "$SIOX" ]] ; then
+	echo "SYNTAX: $0 <SIOX-INSTALL-PATH> [CLEAN]"
+	echo "If clean is set, then ./waf clean install is called"
+	exit 1
+fi
+
+for L in tools/siox-skeleton-builder/layers/*; do
+	NAME=$(basename $L)
+	if [[ ! -d $L ]] ; then
+		continue
+	fi
+	rm $CUR_DIR/$NAME.err 2>/dev/null
 	pushd $L
 	if [[ $CLEAN != "" ]] ; then
 		rm -rf build
 	fi
-	./waf configure --siox=$SIOX --prefix=$SIOX 2>&1
-	./waf install || ERROR="$ERROR $(basename $L)"
+	( 
+	./waf configure --siox=$SIOX --prefix=$SIOX 
+	if [ "$?" != 0 ] ; then
+		touch $CUR_DIR/$NAME.err
+		continue
+	fi
+	./waf install 2>&1 || touch $CUR_DIR/$NAME.err
+	) 2>&1 | tee $CUR_DIR/$NAME.log 
 	popd
 done
-echo $ERROR > error-wrappers.log
-) 2>&1 | tee wrappers.log
 
-ERROR=$(cat error-wrappers.log)
+ERROR=$(ls *.err 2>/dev/null | sed "s/.err//g" || echo "" )
 echo ""
 if [[ $ERROR != "" ]] ; then
-	echo "Error while processing: $ERROR"
-	echo "The configure/compiler output is stored in wrappers.log"
+	echo "Error while processing:"
+	echo $ERROR
+	echo "The configure/compiler output is stored in <LAYER>.log"
+else
+	echo "All wrappers are sucessfully installed!"
 fi
